+++ compiler/driver/tests/import.rs
@@ -1,52 +1,10 @@
 //! Testing under the assumption that `lli` is installed!
 
-use assert_cmd::prelude::*;
+mod runtime_test_support;
+
 use assert_fs::prelude::*;
-use predicates::prelude::*;
-use std::{fs, path::Path, process::Command};
-
-fn compile(
-    file_paths: &[&Path],
-    root_dir: &Path,
-    output_path: &Path,
-) -> anyhow::Result<std::process::Output> {
-    let mut args = file_paths
-        .iter()
-        .map(|p| p.to_string_lossy().to_string())
-        .collect::<Vec<String>>();
-
-    args.push("-o".to_string());
-    args.push(output_path.to_string_lossy().to_string());
-
-    args.push("--root-dir".to_string());
-    args.push(root_dir.to_string_lossy().to_string());
-
-    Ok(Command::cargo_bin(env!("CARGO_BIN_EXE_kaede"))?
-        .args(args)
-        .output()?)
-}
-
-fn compile_project(
-    file_paths: &[&Path],
-    root_dir: &Path,
-) -> anyhow::Result<(assert_fs::NamedTempFile, std::process::Output)> {
-    let exe = assert_fs::NamedTempFile::new("a.out")?;
-    let output = compile(file_paths, root_dir, exe.path())?;
-
-    assert!(
-        output.status.success(),
-        "kaede compile failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    Ok((exe, output))
-}
-
-fn run_compiled_binary(expect: i32, exe_path: &Path) -> anyhow::Result<()> {
-    Command::new(exe_path).assert().code(predicate::eq(expect));
-    Ok(())
-}
+use runtime_test_support::{compile_project, run_binary as run_compiled_binary};
+use std::{fs, path::Path};
 
 fn test(expect: i32, file_paths: &[&Path], root_dir: &Path) -> anyhow::Result<()> {
     let (exe, _) = compile_project(file_paths, root_dir)?;
